Dissolve CborPrinter into Cbor.Ast Encodable and Showable instances
`CborPrinter` was a public object providing two operations — CBOR binary
encoding and RFC 8949 diagnostic-notation rendering — duplicating what
typeclass instances should carry. The binary fold now lives in
`given Cbor.Ast is Encodable in Data` and the diagnostic fold in
`given Cbor.Ast is Showable`, so `.encode` / `.show` are the single routes.
The COSE/CanonicalCbor cross-module callers and the tests move to those.
The unused streaming `emit` variant is dropped.

Inside `object Cbor` the universal `Encodable in Cbor` given is lexically
closer than the new `Encodable in Data`, so the one in-object caller (the
HTTP `abstractable`) and the by-hand bench/test helpers call the instance's
`encoded` method explicitly to stay unambiguous; outside the object plain
`.encode` resolves to the binary instance.
